Move toHexString() to commonMain
