+++ lrucache/src/jvmMain/kotlin/com/mayakapps/lrucache/Utils.kt
@@ -25,16 +25,3 @@ internal fun DataOutputStream.writeString(string: String) = write(string.encodeT
 internal fun DataInputStream.readString() = readString(readByte().toInt())
 internal fun DataInputStream.readString(length: Int) = readBytes(length).decodeToString()
 internal fun DataInputStream.readBytes(count: Int) = ByteArray(count).also { read(it) }
-
-
-internal fun ByteArray.toHexString(): String =
-    buildString(size * 2) {
-        this@toHexString.forEach {
-            val byte = 0xFF and it.toInt()
-            append(HEX_CHAR_ARRAY[byte ushr 4])
-            append(HEX_CHAR_ARRAY[byte and 0x0F])
-        }
-    }
-
-@Suppress("SpellCheckingInspection")
-private val HEX_CHAR_ARRAY = "0123456789ABCDEF".toCharArray()
